Problems 8 and 11 completed
Problem 12 started, however my brute force solution is not optimal, it
is taking so long I have no idea if it is giving correct answer. Its
fine up to 100 divisors or so and then it gets slow. I need to find a
better algorithm to find the total number of divisors.